feat: Add EditUserProfileSection component and integrate it into UserProfileSection
